requirements.txt
streamlit
langchain
chromadb
langchain_community
sentence-transformers
langchain_core
langgraph
langchain_chroma
langchain_openai
pypdf
python-dotenv
openai